Move countByStyle out of StyleManager into StationCounter
StyleManager now exposes only per-station style CRUD plus an entries()
iterator. The presentation-layer aggregation that derives styleId=0 from
totalStations lives next to its sole caller in StationCounter.
