+++ src/frontend/components/StationCounter.test.tsx
@@ -14,14 +14,13 @@ import {
 describe('StationCounter', () => {
     it('should render without errors when map is provided', () => {
         setupGoogleMapsMock();
-        
-        const mockStyleManager = createMockStyleManager({
-            0: 96,
-            1: 2,
-            2: 1,
-            3: 0,
-            4: 1
-        });
+
+        const mockStyleManager = createMockStyleManager([
+            ['001', 1],
+            ['002', 1],
+            ['003', 2],
+            ['004', 4],
+        ]);
 
         const mockStations = createMockStations(100);
         const mockMap = createMockMap();
@@ -31,12 +30,12 @@ describe('StationCounter', () => {
         );
 
         expect(container.firstChild).toBeNull(); // Component renders null
-        expect(mockStyleManager.countByStyle).toHaveBeenCalledWith(100);
+        expect(mockStyleManager.entries).toHaveBeenCalled();
         expect(mockMap.controls[7].push).toHaveBeenCalledTimes(1);
     });
 
     it('should not render when map is null', () => {
-        const mockStyleManager = createMockStyleManager({});
+        const mockStyleManager = createMockStyleManager();
         const mockStations = createMockStations(100);
 
         const { container } = render(
@@ -44,24 +43,29 @@ describe('StationCounter', () => {
         );
 
         expect(container.firstChild).toBeNull();
-        expect(mockStyleManager.countByStyle).not.toHaveBeenCalled();
+        expect(mockStyleManager.entries).not.toHaveBeenCalled();
     });
 
     it('should return null when stations is null', () => {
-        const mockStyleManager = createMockStyleManager({});
+        const mockStyleManager = createMockStyleManager();
 
         const { container } = render(
             <StationCounter styleManager={mockStyleManager} stations={null} styleVersion={0} map={null} />
         );
 
         expect(container.firstChild).toBeNull();
-        expect(mockStyleManager.countByStyle).not.toHaveBeenCalled();
+        expect(mockStyleManager.entries).not.toHaveBeenCalled();
     });
 
     it('should re-render when styleVersion changes', () => {
         setupGoogleMapsMock();
-        
-        const mockStyleManager = createMockStyleManager({ 0: 96, 1: 2, 2: 1, 3: 0, 4: 1 });
+
+        const mockStyleManager = createMockStyleManager([
+            ['001', 1],
+            ['002', 1],
+            ['003', 2],
+            ['004', 4],
+        ]);
         const mockStations = createMockStations(100);
         const mockMap = createMockMap();
 
@@ -69,13 +73,13 @@ describe('StationCounter', () => {
             <StationCounter styleManager={mockStyleManager} stations={mockStations} styleVersion={0} map={mockMap} />
         );
 
-        expect(mockStyleManager.countByStyle).toHaveBeenCalledTimes(1);
+        expect(mockStyleManager.entries).toHaveBeenCalledTimes(1);
 
         // styleVersionを変更して再レンダリング
         rerender(
             <StationCounter styleManager={mockStyleManager} stations={mockStations} styleVersion={1} map={mockMap} />
         );
 
-        expect(mockStyleManager.countByStyle).toHaveBeenCalledTimes(2);
+        expect(mockStyleManager.entries).toHaveBeenCalledTimes(2);
     });
 });
