+++ src/test-utils/test-utils.ts
@@ -98,8 +98,8 @@ export const createMockStations = (count: number, startId: number = 18786) => ({
 });
 
 // Create mock StyleManager
-export const createMockStyleManager = (countByStyleReturnValue: Record<number, number>) => ({
-    countByStyle: vi.fn().mockReturnValue(countByStyleReturnValue),
+export const createMockStyleManager = (entries: Array<[string, number]> = []) => ({
+    entries: vi.fn().mockReturnValue(entries),
     getStyle: vi.fn(),
     changeStyle: vi.fn(),
     resetStyle: vi.fn(),
